New attempt finished lesson.
My first attempt went about doing it incorrectly(I used borders instead
of padding and margin for most of it) took a few days away from this
lesson and tried to see if I could do it diffently. Took me awhile to
figure out the padding on .card. Kept trying to add a margin to title
for some reason(Spoiler, doesn't work that way).